% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ENCODE_PLS_regions}
\alias{ENCODE_PLS_regions}
\title{ENCODE PLS regions}
\source{
{
\url{https://screen.encodeproject.org}
}
}
\usage{
ENCODE_PLS_regions(metadata = FALSE)
}
\arguments{
\item{metadata}{If TRUE, retrieve the AnnotationHub metadata instead of the
object itself. Defaults to FALSE.}
}
\value{
A GRanges object with 40,734 ranges and no metadata.
}
\description{
A GRanges object containing regions of candidate cis-regulatory
elements with promoter-like signatures as identified by the ENCODE SCREEN
project. These consist of regions with high H3K4me3 and DNase signal, and
located within 200 bp of a GENCODE transcription start site. \strong{Citation:}
ENCODE Project Consortium; Moore JE, Purcaro MJ, Pratt HE, et al. Expanded
encyclopaedias of DNA elements in the human and mouse genomes. Nature.
2020 Jul;583(7818):699-710. doi: 10.1038/s41586-020-2493-4. Epub 2020 Jul 29.
Erratum in: Nature. 2022 May;605(7909):E3. PMID: 32728249; PMCID: PMC7410828.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
ENCODE_PLS_regions()
\dontshow{\}) # examplesIf}
}
